+++ MobileBankAppUI/app/src/main/java/com/gmail/matikano9/mobilebankappui/domain/model/Account.kt
@@ -8,13 +8,26 @@ import androidx.compose.material.icons.outlined.MonetizationOn
 import androidx.compose.material.icons.outlined.Savings
 import androidx.compose.ui.graphics.vector.ImageVector
 import com.gmail.matikano9.mobilebankappui.R
+import kotlin.random.Random
 
 data class Account(
     val title: String,
-    val amount: Double,
+    val amount: Double = randomAmount(),
     val type: AccountType,
-    val number: String
-)
+    val number: String = randomNumber()
+){
+    companion object{
+        private fun randomNumber(): String =
+            " ${Random.nextInt(10, 100)} " +
+                    "${Random.nextInt(10000)} " +
+                    "${Random.nextInt(10000)} " +
+                    "${Random.nextInt(10000)} " +
+                    "${Random.nextInt(10000)} " +
+                    "${Random.nextInt(10000)}"
+
+        private fun randomAmount(): Double = Random.nextDouble(1000.0, 99999.99)
+    }
+}
 
 enum class AccountType(
     @DrawableRes val imageRes: Int,