+++ ToDoApp/app/src/main/java/com/gmail/matikano9/todoapp/util/Extensions.kt
@@ -1,5 +1,11 @@
 package com.gmail.matikano9.todoapp.util
 
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.indication
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
@@ -21,4 +27,14 @@ object Extensions {
 
     fun LocalDateTime.toTimeString()
             = this.format(DateTimeFormatter.ofPattern(Constants.TIME_FORMAT))
+
+    inline fun Modifier.noRippleClickable(crossinline onClick: () -> Unit): Modifier =
+        composed {
+            clickable(
+                indication = null,
+                interactionSource = remember { MutableInteractionSource() }
+            ) {
+                onClick()
+            }
+        }
 }